jQuery Content Slider
jQuery is well known for things like image gallerys, sliders, scrollers, etc. We will be building a very clean application which will allow us to click a 'next' button to fade in the next slide and a 'prev' button to go in the reverse direction. We will use the fadeIn() and fadeOut() animation methods to do this