YouTube Search Engine
In this project we will use the YouTube Data API v3 to build a video search engine. We will use the 'search.list' method to do this. We will also implement 'FancyBox', which is a lightbox script. This will allow use to have the video open on our site instead of being redirected to YouTube.